requirements.txt
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.1
blinker==1.9.0
cartesia==2.0.17
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
loguru==0.7.3
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.5
openai==2.9.0
packaging==25.0
pillow==12.0.0
pipecat-ai==0.0.96
propcache==0.4.1
protobuf==6.33.1
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.10.1
pyloudnorm==0.1.1
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
sniffio==1.3.1
sounddevice==0.5.3
soxr==1.0.0
tqdm==4.67.1
twilio==9.8.8
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
python-socketio
